[Refactor] 미션 인증을 통한 레벨 획득 처리 로직 수정

+++ src/main/java/dnd/diary/response/mission/StickerMainResponse.java
@@ -21,10 +21,10 @@ public class StickerMainResponse {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class CurrMissionInfo {
-        private Long subLevel;
+//        private Long subLevel;
+        private Double subLevel;
         private Long mainLevel;
         private Long progressBarRange;
-//        private Long remainToUpMainLevel;   // main 레벨 상승까지 남은 subLevel
     }
 
     @Getter
